Fix 64-bit issues in the Coda client
Use correct RVM layout when unpacking directories from the server.
Make sure we correctly format the directory container file so that it
can be parsed by the kernel module.
Increase the probeslave stacksize to avoid crash.

+++ coda-src/dir/coda_dir.h
@@ -58,9 +58,9 @@ extern int Length (long *);
 /* extern definitions for buffer.c */
 struct buffer
     {
-    long fid[5];	/* Unique cache key + i/o addressing */
-    long page;
-    long accesstime;
+    int fid[5];	/* Unique cache key + i/o addressing */
+    int page;
+    int accesstime;
     struct buffer *hashNext;
     char *data;
     char lockers;
